Core: Move VLog-specific part of InMemoryDataSource to rulewerk-vlog
