+++ rulewerk-vlog/pom.xml
@@ -14,7 +14,7 @@
 	<packaging>jar</packaging>
 
 	<name>Rulewerk VLog Reasoner Support</name>
-	<description> and model</description>
+	<description>Bindings for the VLog reasoner backend.</description>
 
 	<properties>
 		<karmaresearch.vlog.version>1.3.3-snapshot</karmaresearch.vlog.version>
