+++ rulewerk-vlog/src/test/java/org/semanticweb/rulewerk/reasoner/vlog/QueryAnsweringCorrectnessTest.java
@@ -9,9 +9,9 @@ package org.semanticweb.rulewerk.reasoner.vlog;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -71,7 +71,7 @@ public class QueryAnsweringCorrectnessTest {
 	private static final Fact factQg = Expressions.makeFact(predQ, g);
 	private static final Fact factQh = Expressions.makeFact(predQ, h);
 
-	private static final InMemoryDataSource datasource = new InMemoryDataSource(1, 2);
+	private static final InMemoryDataSource datasource = new VLogInMemoryDataSource(1, 2);
 
 	{
 		datasource.addTuple("e");
