+++ rulewerk-vlog/src/test/java/org/semanticweb/rulewerk/reasoner/vlog/VLogDataSourceConfigurationVisitorTest.java
@@ -9,9 +9,9 @@ package org.semanticweb.rulewerk.reasoner.vlog;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -93,7 +93,7 @@ public class VLogDataSourceConfigurationVisitorTest {
 
 	@Test
 	public void visit_InMemoryDataSource_returnsNull() {
-		final InMemoryDataSource inMemoryDataSource = new InMemoryDataSource(1, 1);
+		final InMemoryDataSource inMemoryDataSource = new VLogInMemoryDataSource(1, 1);
 		assertEquals(null, toConfigString(inMemoryDataSource));
 	}
