+++ rulewerk-vlog/src/test/java/org/semanticweb/rulewerk/reasoner/vlog/VLogReasonerWriteInferencesTest.java
@@ -9,9 +9,9 @@ package org.semanticweb.rulewerk.reasoner.vlog;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -80,7 +80,7 @@ public class VLogReasonerWriteInferencesTest {
 	private final Fact f4 = Expressions.makeFact("country", Expressions.makeAbstractConstant("germany"));
 	private final Fact f5 = Expressions.makeFact(universityPred, Expressions.makeAbstractConstant("tudresden"),
 			Expressions.makeAbstractConstant("germany"));
-	private final InMemoryDataSource locations = new InMemoryDataSource(2, 1);
+	private final InMemoryDataSource locations = new VLogInMemoryDataSource(2, 1);
 	private KnowledgeBase kb;
 
 	@Before
